[level 2] Title: [카카오 인턴] 수식 최대화, Time: 1.53 ms, Memory: 77.4 MB -BaekjoonHub

+++ 프로그래머스/2/67257. ［카카오 인턴］ 수식 최대화/［카카오 인턴］ 수식 최대화.java
@@ -59,7 +59,7 @@ class Solution {
         for (i = 1; st.hasMoreTokens(); i++) {
             op = st.nextToken().charAt(0);
             ops[i] = op == PLUS ? 0 : op == MINUS ? 1 : 2;
-            nums[i] = (long) Integer.parseInt(st.nextToken());
+            nums[i] = Integer.parseInt(st.nextToken());
         }
         max = 0L;
         opStack = new ArrayDeque<>();
